visa_guardian/requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2

faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
requests==2.32.3
